Took out the logic of moving a chess piece to a separate class
